bugfix/backend-frontend-images

+++ backend/frankenphp/docker-entrypoint.sh
@@ -1,53 +1,41 @@
 #!/bin/sh
 set -e
+trap 'echo "Error on line $LINENO: $BASH_COMMAND"' ERR
 
 if [ "$1" = 'frankenphp' ] || [ "$1" = 'php' ] || [ "$1" = 'bin/console' ]; then
 	# Install the project the first time PHP is started
 	# After the installation, the following block can be deleted
-	if [ ! -f composer.json ]; then
-		rm -Rf tmp/
-		composer create-project "symfony/skeleton $SYMFONY_VERSION" tmp --stability="$STABILITY" --prefer-dist --no-progress --no-interaction --no-install
+    JWT_DIR="$PROJECT_DIR/config/jwt"
+    JWT_CONFIG="$PROJECT_DIR/config/packages/lexik_jwt_authentication.yaml"
 
-		cd tmp
-		cp -Rp . ..
-		cd -
-		rm -Rf tmp/
+    if [ -n "$JWT_PASSPHRASE" ] && [ ! -f "$JWT_DIR/private.pem" ]; then
+    	echo "Using JWT passphrase: $JWT_PASSPHRASE"
+		echo "Generating JWT keys in $JWT_DIR..."
+		echo "Creating JWT configuration file at $JWT_CONFIG..."
 
-		composer require "php:>=$PHP_VERSION" runtime/frankenphp-symfony
-		composer config --json extra.symfony.docker 'true'
-		composer require lexik/jwt-authentication-bundle
+    	mkdir -p "$JWT_DIR" "$(dirname "$JWT_CONFIG")"
 
-		if grep -q ^DATABASE_URL= .env; then
-			echo 'To finish the installation please press Ctrl+C to stop Docker Compose and run: docker compose up --build -d --wait'
-			sleep infinity
-		fi
-
-	if [ -n "$JWT_PASSPHRASE" ]; then
-		echo "Using JWT passphrase: $JWT_PASSPHRASE"
-
-		JWT_DIR="$PROJECT_DIR/config/jwt"
-		JWT_CONFIG="$PROJECT_DIR/config/packages/lexik_jwt_authentication.yaml"
-
-		mkdir -p "$JWT_DIR" "$(dirname "$JWT_CONFIG")"
-
-        if [ ! -f "$JWT_DIR/private.pem" ]; then
-            openssl genrsa -out "$JWT_DIR/private.pem" -aes256 4096 --passout pass:"$JWT_PASSPHRASE"
-            openssl rsa -pubout -in "$JWT_DIR/private.pem" -out "$JWT_DIR/public.pem" --passin pass:"$JWT_PASSPHRASE"
-            echo "JWT keys generated."
-        fi
+        openssl genrsa -out "$JWT_DIR/private.pem" -aes256 4096 --passout pass:"$JWT_PASSPHRASE"
+        openssl rsa -pubout -in "$JWT_DIR/private.pem" -out "$JWT_DIR/public.pem" --passin pass:"$JWT_PASSPHRASE"
+        echo "JWT keys generated."
 
         if [ ! -f "$JWT_CONFIG" ]; then
-            cat <<EOF > "$JWT_CONFIG"
-        lexik_jwt_authentication:
-            secret_key: '%kernel.project_dir%/config/jwt/private.pem'
-            public_key: '%kernel.project_dir%/config/jwt/public.pem'
-            pass_phrase: '${JWT_PASSPHRASE}'
-            token_ttl: 3600
+             cat <<EOF > "$JWT_CONFIG"
+           lexik_jwt_authentication:
+               secret_key: '%kernel.project_dir%/config/jwt/private.pem'
+               public_key: '%kernel.project_dir%/config/jwt/public.pem'
+               pass_phrase: '${JWT_PASSPHRASE}'
+               token_ttl: 3600
 EOF
-            echo "Lexik JWT configuration created."
+           echo "Lexik JWT configuration created."
         fi
 
-	fi
+    fi
+
+	if grep -q ^DATABASE_URL= .env; then
+    	echo 'To finish the installation please press Ctrl+C to stop Docker Compose and run: docker compose up --build -d --wait'
+    	sleep infinity
+    fi
 
 	if [ -z "$(ls -A 'vendor/' 2>/dev/null)" ]; then
 		composer install --prefer-dist --no-progress --no-interaction
